fix: Resolve CLI invocation and backtest `ValueError`
This commit addresses two core issues:

1.  The CLI was not functioning correctly. Running with `--help` would either trigger a backtest or raise a `TypeError`. This was traced to a known bug in an older version of the `typer` library. The `typer` dependency has been upgraded to `0.16.1` to resolve this. The CLI entrypoint has also been refactored for clarity.

2.  The backtest would crash with a `ValueError` in the `bbands` indicator function. This was caused by an edge case where the input data length was equal to the lookback period. A stricter guard has been added to the function to prevent this calculation and resolve the error.
