+++ pyproject.toml
@@ -14,7 +14,7 @@ yfinance = "^0.2.28"
 pydantic = "^2.0.0"
 python-dotenv = "^1.0.0"
 configparser = "^6.0.0"
-typer = {extras = ["rich"], version = "^0.9.0"}
+typer = "^0.16.1"
 pytest = "^7.4.0"
 mypy = "^1.5.1"
 ollama = "^0.1.7"
