+++ results/backtest_summary.md
@@ -1,0 +1,14 @@
+
+## Backtest Report
+
+**Period:** 2018-01-01 to 2023-01-01
+**Total Trades:** 17
+
+### Key Performance Indicators
+| Metric | Value |
+| --- | --- |
+| Net Annualized Return | -67.26% |
+| Sharpe Ratio | -1.51 |
+| Profit Factor | 0.00 |
+| Maximum Drawdown | -99.62% |
+| Win Rate | 0.00% |
